Stabilize Pyxel 3.0.0 and record design decisions
Separate design principles, audit procedures, and concrete decisions. Apply the reviewed conventions across Rust, Python, Web, documentation, and tests, and synchronize the distributed WASM wheel.

Correct GC reentry, clipping and collision boundaries, and save-failure recovery. The committed tree matches the independently audited candidate; native, Python, and Web checks cover 2,076 passing tests.

+++ crates/pyxel-binding/src/font_wrapper.rs
@@ -12,7 +12,7 @@ impl Font {
             .map_err(pyo3::exceptions::PyException::new_err)
     }
 
-    fn text_width(&self, s: &str) -> i32 {
+    fn text_width(&self, s: &str) -> i64 {
         self.inner_mut().text_width(s)
     }
 }
